Issue #9801: Refactor AutofillCreditCardsAddressesStorage into its own :service-sync-autofill component
