+++ taskcluster/ci/config.yml
@@ -97,6 +97,7 @@ treeherder:
         service-location: service-location
         service-nimbus: service-nimbus
         service-pocket: service-pocket
+        service-sync-autofill: service-sync-autofill
         service-sync-logins: service-sync-logins
         support-android-test: support-android-test
         support-base: support-base
